Fix wording of activity-time-period question
Updated wording based on conversation with Tom Hughes: this question needs to be about the activity checks you can make with an organisation, not about some specific instance of checking something.

We still need to figure out whether this works with GPG45: overall score for "activity" should probably be based on the weakest link (ie lowest score) for _all_ of the organisations from which you get activity history.

https://trello.com/c/GFCeXuJO/186-tweaks-to-activity-journey